requirements.txt
esm>=3.2.0
biopython>=1.85
